Signature lock checking; get the correct field lock specs.
When we are asked to check a signature, we gather the definition
of which fields should be locked. Previously we looked in the
signature to find out which fields the signature should have
locked.

The correct place to look is in the value of the signature when
signed; this allows for signatures that lock different fields to
the definition. This is important because if we check the XFA
to get field locking out when signing the signature, the value
can lock more fields than the signature told us to.

Also, tests with old versions of mupdf that did not fill out
the locking details in the signature value gave us files that
Acrobat was happy to accept as being well formed, and still
allowed fields to be changed without invalidating that signature.
Thus it seems clear that Acrobat must be using the details from
the value, not from the sig.