Avoid multiple incremental sections when saving a signature.
Once the value of a signature field is set, any further changes to the
document cause the creation of a new incremental section. Because of this,
all auxiliary changes associated with a signature must be made before the
field value is set. This commit reorders a case where that rule hadn't
been followed.