Fix bug in fz_compress_ccitt_fax_g4.
Don't return changing color coordinates past the edge of the image.

For a run pair (H white black) where the white run ends at the edge of the
image, we would return an X coordinate past the number of columns in the
image for the black run.

Thus we would encode H runs that went past the edge (H 52 1) instead of
stopping at the edge (H 52 0).