Bug 702475: Handle zlib asking for more memory when finishing deflate.
When deflate() is called with Z_FINISH and returns Z_OK it is asking
for more output space. The caller is then supposed to provide more
space, not change the input data and call deflate() again.

Previously deflate() returning Z_OK was treated as an error. This
commit changes the handling of deflate() returning Z_OK to double the
output buffer size and try again.