Improve diagnostic warnings when store items collide in the hash table.
This happens occasionally when rendering with multiple threads, when
several threads load the same resource and then race to be the first to
insert it into the store.

Now print the type of item we find a duplicate of, instead of relying
only on the fz_hash_table "assert: overwrite hash slot with different
value" message.