java: Allow ongoing search to be aborted or interrupted by UI-events.
This is done by partitioning the search job into a sequence of shorter
jobs that are sequentially inserted into the normal UI interface
eventqueue. By doing this there are no thread synchronization issues
that need to be resolved, except when a search should be aborted. In
this case the invoking thread blocks on a cancellation boolean until
the eventqueue has processed the search runnable, this signals back
to the invoking thread that the search has been aborted.