Protect document page list with FZ_LOCK_ALLOC.
We generally require that document level access only happens
in one thread at a time. Most importantly, this means that only
one thread can ever actually be hitting the actual file at a
time. Another side effect of this is that only one thread can
be modifying the "document structure" - such as the linked list
of loaded pages.

Unfortunately, the java bindings break this, as java can run
finalizers on a background thread. This means we can have a page
being dropped in a background thread due to having no outstanding
references, just as the normal document-accessing foreground thread
tries to reload the page.

We therefore use the ALLOC lock to protect modifications to the
document page list.