Wrap jni_rethrow() and family in macros.
Commit c51cc4c3da519ffdf0fa43d680518ab995332e8b started to
return from JNI functions using comma expressions. Unfortunately
this also meant returning the void return value from jni_rethrow()
in JNI function returning void. Neither gcc nor clang complained
about this (with our normal set of warnings), but MSVC caught it.

Accordingly, we move to using macros for jni_rethrow() and similar
jni_throwing "functions". These wrap up the rethrow and the return
into a single invocation. The values we return here are never
important because the caller should spot that we have raised an
exception before it takes notice of the value we pass back.

Thus, we can have jni_throw_whatever() always end with a return 0,
which will work, whether it's an int, a long, or a pointer expected
as the return value. To cope with throwing in void contexts, however,
we need to also have jni_throw_whatever_void();