Bug 702865: Don't walk through successive pages while linearising.
When preparing a document for linearisation, we walk all the sub
objects to mark those as being used by a given page.

Unfortunately, when we walk annotations, these can have 'Dest'
references that are other pages. The current code will attempt
to walk those too. This means we can recurse a very long way,
causing use to overflow the exception stack.

To avoid this we avoid stepping through references that take us
either to the Page tree or to other pages.