Fix rendering of "stroked" type3 glyphs.
As seen with fts_23_2312.pdf from the PDF 2.0 FTS.

This file uses "2 Tr" (Fill and Stroke text rendering mode) to
render a type 3 glyph. The fill part renders the glyph to a pixmap
that goes into the cache. The stroke part then finds the same glyph
and renders that too.

The difference is that the fill code in the draw device understands
how to draw from pixmaps, where the stroke code does not. Fix that
difference here.