Add JBIG2 support to pdf_add_image.
We need to parse the segment headers in order to rewrite the JBIG2 data
to convert it to a PDF compatible JBIG2Decode stream.

1) Strip the file header (if present).
2) Copy segments from 'globals' symbol table (if present).
3) Convert random access format to sequential access (if necessary).
4) Delete end-of-page and end-of-file segments.
5) Delete all segments related to other pages than the first.

TODO: Copy a given page from a multi-page JBIG2 stream. We currently only
support creating a PDF image from the first page of a JBIG2 file.