Add USE_TESSERACT and USE_EXTRACT make variables.
If USE_TESSERACT=yes then we'll build the Tesseract submodule (or use
the system one if USE_SYS_TESSERACT=yes) and enable its use. This will
also add a suffix to the output build directory, so builds with and
without Tesseract enabled can co-exist.

Likewise for the (so far experimental) Extract library.