Update parent object for linked objects when saving/repairing.
Not doing this causes a problem when saving a PDF where a single
signature has been signed more than once.

Before such a PDF is saved to file, each signature's signature
dictionary contains default values for the /ByteRange and /Contents.
Later when the file is saved to disk, first all objects are written to
file (including all signature dictionaries with default values for
/ByteRange and /Contents). Next complete_signatures() needs to update
those default values to the correct values: The /ByteRange value needs
to reflect the correct ranges of bytes used to compute the signature
digest, and the /Contents value must be updated to contain the
computed digest. To know at what file offsets to patch in these
updated values, complete_signatures() queries /ByteRange for its
parent object, expecting the object number of the signature
dictionary, and then queiries the xref for that object's file offset.
This file offset is used to looking for a PDF signature dictionary
that contains the /ByteRange and /Contents in need of patching.

complete_signatures() not patches /ByteRanges and /Contents in the
file, it also updates the in-memory signature dictionary PDF objects
containing the /ByteRange and /Contents. Updating a key-value pair in
a PDF dictionary will normally cause the value's parent object number
to be set to that of the dictionary containing the value. This doesn't
happen when complete_signatures() updates /ByteRange and /Contents.
The reason is that pdf_save_document() calls prepare_for_save() which
sets the flag save_in_progress on the document. This flag prevents,
e.g. pdf_dict_put() and friends from, setting the parent object number
of dictionary values when they are updated. When pdf_save_document()
later calls do_pdf_save_document() which calls complete_signatures()
it creates new PDF arrays and hexstrings their parent object number is
0. These are then used to update /ByteRange and /Contents in the
in-memory PDF dictionaries, but the value's parent object number
remains at 0.

When only saving a PDF with a single signed signature this is not a
problem, because the first time /ByteRange is queried for its parent
object number by complete_signature() it contains object number set
when the default value for /ByteRange was inserted into the dictionary
during signing of the signature.

When saving a PDF with a signature that has been signed more than once
there is a problem. As complete_signatures() is processing the first
signing of the signature it asks /ByteRange for its parent object
number, gets the object number of the signature dictionary, which in
the xref points to a proper file offset and the computed byte range
and signature digest can be patched into the signature dictionary's
/ByteRange and /Contents values in the file. As complete_signatures()
is processing the second signing of the same signature it again asks
/ByteRange for its parent object number, it unexpectedly gets 0 (since
the /ByteRange value was updated in-memory during the completion of
the first signature), the xref is then asked for the file position of
PDF object 0, which is located at file offset 0. complete_signatures()
then starts to look for a signature dictionary in need of patching at
file offset 0, but in the end errors out because it can't find the
signature dictionary.

This commit allows newly linked array and dictionary objects to have
their parent object numbers be updated even when a PDF file is being
saved (or repaired). This fixes the issue in the previous paragraph
by allowing /ByteRange's parent object number to be updated to that of
the signature dictionary containing it. This leads to
complete_signatures() being able to locate the signature dictionary's
file offset and patch its values.