Bug 702970: Fix clipping problems with group alpha.
If we have a non-rectangular clipping region, in the presence of
group alpha, we make a new group alpha plane (as we need to, so
we don't render to clipped areas), but initialise it to be empty.
We ought to copy it from the existing group alpha plane.

Same deal for shape planes.