Coverity 177127: Fix fz_copy_pixmap_area_converting_seps
In the case where we are copying a pixmap area and changing
colorspaces and our source colorspace is a NON device-n one
with spots with some of the spots set to be disabled or unnamed,
then we might have been setting up the map array incorrectly.

The new handling is simpler, and matches the revised comments
in the code.