Coverity 173399 et al: Avoid sign extension problems.
Coverity sees problems in the top 32 bits of 64bit "longs" with AES.
The code we have was written for 32bit longs, and works for 64 bit
longs purely by the fact that the top 32bits aren't accessed.

To assuage Coverity's sensibilities (and to be neater overall) we
move to using uint32_ts rather than "unsigned longs", so we are
consistent on all platforms.