Bug 703092: Fix logic error in structured text scanning.
When scanning the MCID string for a matching character when running
the process filter, in the case where we do not remove the character
we could get stuck at a space character in the input since we'd not
progress.

When we match a character we should either remove or skip it, so make
that decision more explicit.

This also fixes another potential bug where when we were removing
a matching character, str->pos would also be advanced so we would miss
scanning the next character after it.