Coverity 215306: Limit field size.
In pdf_parse_jbig2_segment_header() Coverity determines that
the data variable must come from an untrusted source since it
is byte swapped before being used. This means that any value
read from the data varible must have its upper and lower
bounds tested explicitly.

In particular this issue points to the 3 bit rts value only
having its lower bound tested, but not its upper bound. We do
read the unsigned char data and shifts the value 5 bits to
the right, which effectively ought to limit the value to the
range [0,7]. It is not quite clear why Coverity can't see
this, or where we might be mistaken.

So here we try to mask out the 3 necessary bits after having
done the shifting to clearly tell Coverity that we want to
limit the upper bound to 7.