Bug 703219: Handle short skips in the subarea stream.
The subarea stream will skip over a number of bytes before it
starts returning bytes from the underlying stream. Previously
when the underlying stream did not successfully skip over the
desired number of bytes subarea_next would return 0, which is not
related to the value of the first byte of the underlying stream.

The subarea stream has now been changed so that it, instead of
returning 0, will retry skipping the remaining number of bytes.
This repeats until all bytes have been skipped, or the underlying
stream failed to skip over any bytes at all. The latter case is
treated as if the underlying stream returned EOF, hence the
subarea stream itself also returns EOF.