Bug 703100: Avoid encrypting (and extending) signature digests
When creating a new digital signature in a PDF MuPDF asks the
PKCS7 signer for the maximum digest size and adds a zeroed out
dummy signature as a /Contents string to the signature
dictionary. When MUPDF saves the document to a PDF file using
encryption it used to also encrypt the /Contents entry. For
AESv2/v3 encryption schemes strings will be extended by 16-31
bytes owing to an preceding iv string and trailing padding. Next
MuPDF completes unsaved signatures by locating their signature
dictionary objects in the PDF file, finding the /Contents entries
and overwriting their values with the an unencrypted signature
digests calculated by the PKCS7 signer according to the ranges
specified by each signatures /ByteRange entry.

For the PDF from the bug the initial zeroed out signature digest
was extended by 17 bytes when encrypted, thus the reserved size
for the /Contents entry was 17 bytes larger. When completing the
unsaved signatures MuPDF would previously, in pdf_write_digest(),
compare this reserved size to that of the unencrypted zeroed out
dummy signature. These differed by 17 bytes and
pdf_write_digest() errored out.

The check revealed that the zeroed out dummy signature digests
were written using encrypting and then patched without
encryption. Despite this the check is misguided because it
compares the encrypted and unencrypted lengths of the zeroed out
dummy signature digest. The check was meant to compare the
reserved space for the signature digest to the size of the digest
calculated by the PKCS7 signer.

This commit implements the correct approach which is to not
encrypt the initial dummy signature digest /Contents entry while
writing objects to the PDF file, and compare the size of the
digest returned by the PKCS7 signer to the reserved space in the
PDF file. The check has been modified to throw an exception if no
digest is received or if the digest is larger than the reserved
space. Should the digest be shorter than the reserved space,
it will be zeropadded at the end until it cover the entire space.