Remove pdf_annot->ap.
No longer hold the 'current' annotation appearance as a ref in the
annot. Instead, fetch it when required.

Firstly, this is in line with our current approach of trying to
minimise the information held outside of the document.

Secondly, when we move to doing appearance synthesis outside the
main document (using "local" xrefs), this saves us having to cope
with moving the cached ap value in/out of the local xref too.