Add local_xref mechanism for PDF docs.
We introduce the idea of a local_xref, to hold changes to the
document that exist purely because of having synthesised appearances
for annotations that do not have appearance streams. We do not want
such synthesised appearance streams to be written back into the
document itself, unless we have edited the annotation.

Thus, every pdf_document has a local_xref pointer to hold such
changes. We arrange for this to be "pushed" before each read or
synthesis operation on an annotation, and "popped" afterwards
(these operations resolve to simply incrementing/decrementing
a "local_xref_nesting" count in pdf_document). Whenever we both
have a local_xref pointer and a non-zero nesting, we hold the
local_xref to be in force.

Whenever we go to update the appearance of an annotation
(pdf_update_appearance), we push the local_xref. We figure out if
a new appearance stream is required because there isn't one (in
which case we should generate one locally) or because the annotation
has been edited. In the former case, we leave the local_xref in place
and generate stuff into the local_xref. In the latter case, we
throw away any local_xref and generate our appearance into the document
proper.

We push/pop the local_xref around all read operations on an annotation,
so we read/write things like the Rect and appearance stream from
our newly synthesised versions, if any.

Edits to the document fall into 2 cases. If the edit comes from a
caller to mupdf making a real change to a document (say setting
the rectangle), then the local_xref will not be pushed. When we
reach pdf_prepare_object_for_alteration, we therefore discard the
local_xref before making the change. This is therefore a real and
lasting change as we want.

If an edit to the document is caused as a part of annotation
synthesis, then the xref will be in force by the time we get to
pdf_prepare_object_for_alteration, so any change will go to the
local_xref (and thus not cause a write to the REAL document).

Thus any editing to the document OTHER than in the synthesis case
first discards the local_xref, so we operate directly on what is
in the document and the edits are permanent.