Support both version 4.x and 5.x of Tesseract.
Tesseract 4.x is the stable version while MuPDF bundles 5.x.
This means that when using the system versions of Tesseract it
is likely MuPDF is being linked to some version of Tesseract 4.x.
This commits adds support for compiling with both versions.
Only MuPDF compiled for desktop Java has been linked with
Tesseract 4.x on Linux, but it appears to work fine.