Link to stdc++ even when using system libraries.
The reason this is needed is because tessocr.cpp calls new/delete/etc.
which come in libstdc++. When building MuPDF's bundled tesseract into
binary TESSERACT_LIBS explicitly adds -lstdc++ since tesseract itself
also need the same library. Thus the dependency from tessocr.cpp on
libstdc++ has been invisible. When linking with the system library
version of tesseract on the other hand its dependency on libstdc++ is
not carried over when linking in tessocr.cpp. pkg-config --libs for
tesseract does not list -lstdc++ since any user ought to list their
own dependencies. And with this commit, this is exactly what we do.