Add Snapshot option to PDF saving.
The purpose of a "snapshot" save is to allow us to dump the current
state of a PDF document (including edits, but excluding undo/redo
history) in such a way that the version in memory remains unchanged.

There are a couple of use cases for this:

1) Load a form, fill in some fields, print it. In order to do the
print, we need to save the document as a standard valid PDF to send
to a remote print service. After printing, if we then edit the
document some more and save it out, we only want to see 1 incremental
section used, rather than 2 (i.e. the saving for printing should not
cause the 'underlying' document to be updated).

2) When running as an app on a mobile device, when we are put into
the background, we need to save our state so that if the app is killed
and later restarted, we can pick up where we left off. Again this
should not involved writing a new incremental section to the document.

This commit solves for case #1.

Case #2 will require this, plus both the ability to save undo/redo
history, and the ability to 'reopen' the last incremental update.