Fix 'wonky bboxes' issue in OCRd text.
When we save a 'High Security Redaction' version of a document
(even without any redactions in)  the resultant file often
gives 'wonky' selection boxes when viewed in Acrobat (or similar).

This is because the OCR routine will return different word bboxes
for different words, and baselines won't line up.

Accordingly, we now collect the recognised words up as lines before
flushing them to the stream, and use a line bbox rather than a
word bbox to base the positions on.

This code currently relies on text being horizontal. Vertical mode
text will probably break it. The existing code relies on both
L2R and horizontal text, so we are no worse off with this.

We can cope with R2L and vertical text when we have some examples.