Tweak pdf_update_stream to work better with undo.
Set the new length in the object before we update the stream.
Setting the new length has the effect of moving/copying the
old object/old stream into the journal. Previously we were
setting the stream, THEN moving the old version which meant
the new stream was copied in to accompany the old object.