Bug 701640: Fix linearisation offset problem.
When linearising, we write objects out in one pass to establish offsets.
We then create the hint objects, and do another pass to properly
write the file out.

During the first pass, we were creating a hint object without a
stream. The stream was then created before the second pass was
performed. Accordingly, the overall length of that object changes.
The code allowed for the change in the size of the stream, but
was failing to allow for it changing from not having a stream at all
to having a stream. (i.e. it allowed for the stream changing size,
but did not allow for the addition of "stream\n" and "endstream\n".)

Due to the 'slack' inherent in writing objects with default large
numbers in the first pass, and those numbers being reduced to smaller
values in the second pass, we were getting away with this in most
cases. The addition of an ASCIIHexDecode filter was tipping us over
the edge.

The fix is simply to ensure we create the object with a stream to
start with.