source/fitz/output-docx.c: updated to match changes to extract library.
We now create an extract_alloc_t* that uses fz_realloc_no_throw() as the
underlying allocator, and pass it to all extract_*() functions.

Improved how we store the fz_context* in fz_docx_writer, to ensure that things
work even if we are called (non-concurrently) on different threads.

Updated submodule thirdparty/extract.