Bug 702261: Tweak resolving of OCGs.
OCGs are defined as indirect objects with a name. OCGs are similarly
referenced as indirect objects.

Previously, when comparing OCGs we resolved the indirect objects
and compared the contents of the objects. As this bug shows, however,
we can have 2 different OCGs with the same name, but different states.

Accordingly, we now compare the indirect objects directly, without
resolving the objects, and so only match if the object is identical,
not if the objects merely happen to have the same names.