Generate signature appearances via the device interface
Generate a display list via the device interfaace and update the appearance
using pdf_update_appearance_from_display_list. Using this mechanism is
likely slower but provides an easy route for appearances of a more
general form.

As part of this change, the hard limit of 5 lines has been dropped for the
certificate info, so that there is now no risk of some part overflowing
and not being visible. Also the mupdf logo has been used for the water
mark, rather than a dingbats glyph.