Provide more flexibility in the generation of signature appearance streams
Add pdf_sign_signature_with_appearance to allow an arbitrary display list
to be the source of the appearance stream.

Add pdf_signature_appearance to construct a display list for the common
layouts of text or image on the left and text on the right, with an
underlying logo. If nothing is specified for one side then what is
specified for the other fills the whole rectangle.

Add pdf_signature_info to construct formatted text in the form often used
for the righthand side of a signature appearance.

Remove pdf_update_signature_appearance and
pdf_update_signature_appearance_with_image, since they are no longer
needed.