Fix several more similar ocurrences of Bug 703666.
Bug 703666 exhibited a leak of page resources and a softmask in
case of an exception. The fix for that bug resolved the issue in
pdf_show_path(), but the same type of leak was present in
pdf_show_shade(), pdf_show_image() and pdf_flush_text(). These
leaks are all fixed by this commit.