scripts/jlib.py: optionally specify -rpath in link flags. other minor fixes.
link_l_flags(): On Linux we append " -Wl,-rpath='$ORIGIN'" to link commands
which allows .so's to be found at runtime when they are in the same
directory. This appears to be unnecessary on OpenBSD.

Added new functions remove_dir_contents() and ensure_empty_dir().

Simplified force_line_buffering() - we don't need os.dup().

Fixed double prefix of log lines.