Bug 703753: Add detection of Orientation from Exif images.
Add the concept of 'orientation' to fz_images. This is set to
the orientation defined in an image; currently only Exif format
images set this. This value is ignored by all internal rendering
functions.

The image document handler checks the orientation, and flips/rotates
the image as necessary when feeding it to the display functions.