Use a typedef rather than an ifdef for fz_jmp_buf.
This is neater. Unneccesary source transformations are bad.

This was spotted with the C++ wrappers; when generated on
Linux, and compiled on Windows, they ended up using sigjmp_buf
rather than fz_jmp_buf, which lead to a compile failure.

I believe Julian is fixing the generation to avoid the relevant
functions at generation time, but having spotted the problem
best to fix it.