Bug 703756: Update alpha state in pdf_dev_alpha.
We were checking the new alpha value with the old one, but we never
updated the old one.

In this case, the file starts by drawing with alpha of 0, and that
would emit an alpha change because it's different from the initial
value of 1.

Then the file would draw with alpha of 1, but since we believe we're
already at alpha 1, we don't change the alpha and keep emitting graphics
with alpha 0.