Update ensure_initial_incremental_contents to take the expected length.
This will allow us to cope with doing 'incremental' writes to/from
files that have extra information at the end that should be dropped.
This will be useful when we come to write incrementally from files
that have been 'snapshotted' and we want to write again while
dropping the last (transient) xref on those files.