Bug 703873: Workaround curl issue causing SEGV.
Curl needs to be told to treat http errors as being fatal.

We fetch the curl data in 2 lumps; first the header, then the body.
In the event that the header returns 405, curl was returning no error to
the caller, but then SEGVs when asked to fetch the data.

Thanks to Sebastian for pointing out the curl option to make it
fail on error.