android: Enable SIMD compilation flags per target ABI.
When Tesseract was enabled this used to fail at link time because on ARM64
thirdparty/tesseract/src/arch/simddetect.cpp refers to intSimdMatrixNEON
which was not provided. Tesseract's cmakefile provides settings for SIMD
extensions such as NEON, AVX, AVX2, FMA and SSE4.1. MuPDF does not use
cmake, but instead checks the make variables HAVE_NEON, HAVE_AVX, etc. in
Makelists and enables corresponding CFLAGS for Tesseract. The problem was
that these make variables were never set. This commit sets the flags
in Android per target ABI before including Makelists, thereby ensuring
that intSimdMatrixNEON is provided on ARM64 which causes the linking to
succeed. AVX, AVX2, FMA and SSE4.1 are enabled for Android x86-64.