Avoid checking color when writing Default appearance.
Previously MuPDF checked whether the color was black when generating
an annotation's Default appearance. The idea being that black color
in RGB can be expressed in gray colorspace instead.

First, the check for black color for CMYK was incorrectly written (since
K==1 for black in CMYK, not 0).

Second, black color is translated differently from gray and RGB
colorspaces into CMYK when ICC is enabled (see section 6.3 in pdfref17
and fz_init_process_color_converter()). This means that the colorspaces
are not equivalent and so the number of components should not be
optimized in this way.

Third, if a murun script tried to set the color [0, 0, 0] using
PDFAnnotation.setDefaultAppearance() and then retrieve it again using
PDFAnnotation.getDefaultAppearance() it would get the color back as [0].
This confused the murun test scripts as they expected the same number of
components out as it provided in the first call.

With this patch the number of components are preserved and there is no
optimization of how to express the color black.