Indicate the number of components when setting text color.
pdf_field_set_text_color() first parses the default appearance to remember
font name and font size, then processes the provided text color, before
finally setting the complete default appearance at the end. There was a bug
where the code expected the number of components in the original and new
text color to be the same, but they might not be. So ensure that the number
of components in the new text color is used when setting the new default
appearance (font name, font size and text color).