Add support for creating pixmap from float array using tone-mapping.
Tone-mapping is necessary to convert high dynamic range float samples
into low dynamic range 8-bit unsigned integer samples with good results.

The paper "Which tone-mapping operator is the best? A comparative study
of perceptual quality" by C. Alejandro Parraga and Xavier Otazu surveys
a number of tone-mapping algorithms. The paper concludes that one of the
better algorithms that doesn't underperform in any metric is an
algorithm described in the 2008 paper "Consistent tone reproduction" by
Min H. Kim and Jan Kautz.

This commit implements the Kim/Kautz algorithm as part of creating a
fz_pixmap from an array of float samples.