Estimate negative stream lengths as being zero.
PDF object 5 in tests_private/pdf/customer394/problems/normal_994.pdf
has a /Length field of -1121. In pdf_load_image_stream() we propagate
this signed integer value to an unsigned size_t and it gets
interpreted as a large value, causing a memory allocation to fail and
the PDF not to be rendered.

This commit adapts the stream length guessing function to assume that
negative lengths are supposed to be zero. This guess is probably not
the correct stream length, so warnings about incorrect stream lengths
are likely. After this commit MuPDF can parse and render normal_994.pdf
with warnings.