Fix Leptonica to use MuPDF memory allocators.
A typo in the Leptonica build options was causing Leptonica to use
native malloc/free on non-windows platforms.

Tesseract allocates via new/delete/new[]/delete[]. While these
can be trapped (see tessocr.cpp in the Ghostscript source for
an example of how), this does not help us here.

Tesseract calls new/delete etc as part of the static init it does
on startup. This means that any redirection of new/delete to
other allocators would have to happen before MuPDF starts up. This
does not fit with MuPDF's method of having the allocators defined
in the context.

We therefore live with Tesseract using the standard allocators
(new/delete). If this really is a problem for any particular
integrator, then we work with them to hardwire in allocator changes.