Tweak where we gridfit images, so it happens before subarea.
Currently, we find any subarea of an image we can restrict our
rendering too, decode that area, and then grid fit that reduced
area. This means that rendering in tiles, we might get different
results to rendering the whole page - something that really ought
not to happen.

Move the gridfitting out of fz_paint_image (and
fz_paint_image_with_color) into the callers, before the subarea
calculations are done.

This restores sanity.