Avoid stray objects in incremental xref after pdf_delete_object.
When we create an object, it is put into the incremental xref.
If we delete that object, then a 'free' object is overwritten
in its place.

We have it in mind to move to spotting objects in the incremental
xref as meaning 'we have unsaved changes', and this is scuppered
by having needless 'free' objects in there.

Objects that existed in previous versions of the document and have
now been removed, certainly need to have 'free' object entries
(and these should count as changes). Objects that have never
existed until this version should be deleted entirely, by being
given a 0 type (and hence should not count as changes).

Accordingly in pdf_delete_object, we check back through the
previous xrefs; if the most recent entry for this object is free
(or if there is no reference at all), then we know we can delete
it by removing the entry entirely.