Tweak trailer object handling for snapshotting.
When we read a document in, if it has a new format xref, we read
the trailer object and insert it into the xref entry. We have
(until now) forgotten to set the parent number for that object.
Do so.

When we read a journal in, if we 'reopen' the incremental section
of the document, remove the trailer object we read from the xref.
This will be recreated anyway on any future save.

The advantage of doing this is that it puts us back into the same
state as we would be had we never snapshotted and reloaded (which
is, after all, the purpose of snapshotting).

This sets us up nicely to be able to "undo" back to a state of
"no changes" (based upon the contents of the incremental xref).