Bug 704045: Purge local_xref keyed objects from store on local_xref destruction.
This avoids us being left with fonts in the store keyed on
objects that no longer exist.

Part of this commit is that we make pdf_is_local_object respond
whether the xref is in force or not. The current place it is
called from always has the local xref in force, so this doesn't
alter any behaviour.