Update pdf_update_page to cope with local xrefs.
pdf_update_page runs across all the annotations on a page
calling pdf_update_annot. This in turn calls pdf_update_appearance.

A call to pdf_update_appearance can either a) do nothing,
b) synthesise an appearance in a local_xref or c) synthesise
an appearance in the real document, discarding any local_xref.

Thus a call to update the appearance of an annotation that does c)
can 'undo' the update of earlier annotations.

The fix for this is to spot the change, and to rerun the updates
for earlier annotations. This requirement is now documented in
the header file.

We update the implementation of pdf_update_page to perform this
correctly.

We also take the opportunity to hide pdf_update_appearance from
the API, as people should just be calling pdf_update_annot.