pdf_create_annot: return an owned reference not a borrowed one.
This seems much more logical - indeed, the Java, C++ and Python
bindings have all been written to assume this.

Knock on effects are for pdf_annot_create_raw and
pdf_create_signature_widget to also return owned references.

mupdf-gl has been updated to reflect the fact that it owns the
reference.

fz_create_link and pdf_create_link are similarly updated. The
java (and presumably C++ and python) handlers for this already
assumed that they returned an owned reference.