pdfwrite: Support substitute CJK fonts in output.
Track CJK fonts when creating them and store the language in the font
flags. Use the language flag to create substitute CJK fonts with
the appropriate character collection encoding.

Use UCS2 CMaps for the text encoding when creating a substitute font.

The font metrics of the substitute font may not always match, but some
text is better than no text!

Also add in a generic substitute font creation stub.