Fix SEGV in pdf_add_journal_fragment.
With specific files we can get SEGVs in pdf_add_journal_fragment
on loading the document. We are arriving there with no journal
entries defined.

This is because we've hit a PDF repair, which has discarded the
journal, before we've continued though to make a change.

We fix this in two ways.

Firstly, we no longer discard the journal on a repair. The journal
should not be dependent on anything that can change during a repair,
so this should be safe. This should be nicer for a user, because
losing your undo history due to a repair triggered late on in
using a file is unhelpful.

Secondly, we introduce the ability to tell the core to 'throw after
repair'. For certain sections of the code (here, the annotation
synthesis code), if we get a repair during the code, we want to
perform the repair, then restart the operation from scratch.

Accordingly, we have start/end functions to trigger this behaviour,
changes to pdf_repair so that we throw as required, a new type of
error so we can recognise this case, tweaks to the catch/rethrow
routines used within the appearance synthesis, and top level changes
in the appearance synthesis routines to start/stop and restart as
required.

fix